requirements.txt
streamlit
pandas
numpy
sentence-transformers
scikit-learn
PyPDF2
pdfplumber
nltk
python-dotenv
